+++ CHANGELOG.md
@@ -25,6 +25,7 @@
   * Message queues: handlers now receive message id along with other args.
   * Message queues: `enqueue` now takes an optional flag to specify dedupe behaviour when message is locked.
   * Message queues: `queue-status` now shows next & last mid.
+  * Message queues: add support for delayed start (use `:auto-start <ms>` option).
 
   * Tundra: completely refactored design for robustness+simplicity - now use standard message queue for replication worker.
   * Tundra: replication workers have inherited message queue features like retries, backoffs, etc.